created service_manager which manages all the services (i.e. subscriber, pusher, puller), put puller in it's own thread, cleaned up logging, other performance improvements
