+++ env_activate.cmd
@@ -1,0 +1,1 @@
+call "C:\PYTHON\Scripts\activate" zeromq-high-speed-subscribers
